+++ src/index.js
@@ -1,27 +1,12 @@
-/* eslint-disable import/no-cycle */
 import './style.css';
+import moreIcon from './more.svg';
 import Icon from './enter.svg';
 import recycle from './recycle.svg';
-import { display } from './status';
+import elementGenerator from './status';
 
-export const elementGenerator = (typeName, className, content, idName) => {
-  const element = document.createElement(typeName);
-  if (className) {
-    element.className = className;
-  }
-  if (content) {
-    element.textContent = content;
-  }
-  if (idName) {
-    element.id = idName;
-  }
-  return element;
-};
-
-// eslint-disable-next-line import/no-mutable-exports
-export let todoTasks = [
+let todoTasks = [
   {
-    description: 'Solve one coding challenge',
+    description: 'Read the last chapiter of Ruby book',
     completed: false,
     id: 0,
     checked: false,
@@ -39,6 +24,10 @@ export let todoTasks = [
     checked: false,
   },
 ];
+
+const savedList = () => {
+  localStorage.setItem('ToDo', JSON.stringify(todoTasks));
+};
 
 const todo = elementGenerator('div', 'container', null, null);
 const todoHeader = elementGenerator('div', 'title', null, null);
@@ -62,7 +51,7 @@ enterIcon.classList.add('enter-icon');
 
 form.appendChild(enterIcon);
 
-export const todoList = elementGenerator('ul', 'to-do-list', null, null);
+const todoList = elementGenerator('ul', 'to-do-list', null, null);
 
 const divClear = elementGenerator('div', 'div-clear', null, null);
 const btnClear = elementGenerator('button', 'clear', null, null);
@@ -77,6 +66,58 @@ todo.appendChild(divClear);
 
 const toDoContainer = document.getElementById('todo-container');
 toDoContainer.appendChild(todo);
+
+function display() {
+  todoList.innerHTML = '';
+
+  function listItem(elem) {
+    const list = elementGenerator('li', 'task', null, null);
+    const flex = elementGenerator('div', 'flex', null, null);
+    const oneTodo = elementGenerator('input', 'one-todo', null, null);
+    oneTodo.type = 'checkbox';
+    oneTodo.checked = elem.checked;
+    const form = elementGenerator('form', 'edit', null, null);
+    const input = elementGenerator('input', 'label', null, null);
+    input.setAttribute('name', elem.id);
+    const image = elementGenerator('img', 'more', null, null);
+    image.src = moreIcon;
+
+    input.addEventListener('blur', (e) => {
+      image.src = moreIcon;
+      e.preventDefault();
+    });
+    input.value = elem.description;
+
+    function statusUpdate(item, input, oneTodo, todoTasks) {
+      if (item.checked) {
+        input.classList.add('line-through');
+      } else {
+        input.classList.remove('line-through');
+      }
+      oneTodo.addEventListener('change', () => {
+        const todo = todoTasks[item.id];
+        todo.checked = !item.checked;
+        todo.completed = !item.completed;
+        display();
+      });
+    }
+
+    statusUpdate(elem, input, oneTodo, todoTasks);
+    form.appendChild(input);
+    flex.appendChild(oneTodo);
+    flex.appendChild(form);
+    list.appendChild(flex);
+    list.appendChild(image);
+    todoList.appendChild(list);
+    return list;
+  }
+
+  todoTasks.forEach((elem) => {
+    const item = listItem(elem);
+    todoList.appendChild(item);
+  });
+  savedList();
+}
 
 window.addEventListener('load', () => {
   const result = localStorage.getItem('ToDo');
